7.1.27

+++ include/rain/networking/tcp/socket.hpp
@@ -137,7 +137,7 @@ namespace Rain::Networking::Tcp {
 			virtual int sync() noexcept override {
 				// Send available buffer.
 				try {
-					std::size_t bytesSent{0}, bytesTotal{this->pptr() - this->pbase()};
+					std::size_t bytesSent{0}, bytesTotal{static_cast<std::size_t>(this->pptr() - this->pbase())};
 					while (bytesSent < bytesTotal) {
 						auto result = this->socket->send(
 							this->sendBuffer + bytesSent,
